requirements.txt
whisper
openai-whisper
torch
torchaudio
transformers
google-generativeai
ffmpeg-python
numpy
scipy
requests
urllib3
